added some animation

+++ ClientSide/Entities/Characters/Races/race.h
@@ -8,7 +8,7 @@ class race {
 		int maxManaIncrease = 0;
 		int maxHealthIncrease = 0;
 		int maxStaminaIncrease = 0;
-		int speed = 30;
+		float speed = 30;
 		std::string raceName;
 		//resistances/weaknesses
 };
